+++ src/main/java/cartelera/services/impl/CinemaServiceImpl.java
@@ -10,9 +10,11 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import static cartelera.utils.Utils.invalidPosNumber;
+
 @Slf4j
 @AllArgsConstructor
 @Service
@@ -30,32 +32,37 @@ public class CinemaServiceImpl implements ICinemaService {
 
     @Override
     public Optional<Cinema> findById(Long id) {
+        log.info("findById {}", id);
+        if (invalidPosNumber(id)) return Optional.empty();
         return cinemaRepo.findById(id);
     }
 
     @Override
+    public boolean existsById(Long id) {
+        log.info("existsById {}", id);
+        if (invalidPosNumber(id)) return false;
+        return cinemaRepo.existsById(id);
+    }
+
+    @Override
     public Cinema save(Cinema cinema) {
+        log.info("save {}", cinema);
         return cinemaRepo.save(cinema);
     }
 
     @Override
     public void deleteById(Long id) {
+        log.info("deleteById {}", id);
+
+        if (invalidPosNumber(id) && !existsById(id)) return;
 
         // borrar todas las rooms asociadas
         List<Room> rooms = roomService.findAllByCinemaId(id);
-        List<Long> ids = new ArrayList<>();
-        for (Room room : rooms)
-            ids.add(room.getId());
-        roomService.deleteAllById(ids);
+        if (!rooms.isEmpty()) for (Room room : rooms) roomService.deleteById(room.getId());
 
-        // borrar address asociada
-        Optional<Cinema> cinemaOpt = findById(id);
-        if (cinemaOpt.isPresent()) {
-            Cinema cinema = cinemaOpt.get();
-            Long addressId = cinema.getAddress().getId();
-            cinema.setAddress(null);
-            addressService.deleteById(addressId);
-        }
+        // desasociar la dirección asociada
+        Long addressId = cinemaRepo.findById(id).get().getAddress().getId();
+        if (addressService.existsById(addressId)) addressService.deleteById(addressId);
 
         cinemaRepo.deleteById(id);
     }
